refactor(mediawiki): extract ratelimit, security, http server from main.go (deslop #59)
