+++ main_ratelimit.go
@@ -1,0 +1,91 @@
+package main
+
+import (
+	"sync"
+	"time"
+)
+
+// RateLimiter implements a simple token bucket rate limiter per IP
+type RateLimiter struct {
+	mu       sync.Mutex
+	clients  map[string]*clientLimiter
+	rate     int           // requests per interval
+	interval time.Duration // interval duration
+	cleanup  time.Duration // cleanup old entries
+	stopCh   chan struct{} // graceful shutdown signal
+	stopOnce sync.Once     // ensure single shutdown
+}
+
+type clientLimiter struct {
+	tokens    int
+	lastCheck time.Time
+}
+
+// NewRateLimiter creates a rate limiter with specified rate per interval
+func NewRateLimiter(rate int, interval time.Duration) *RateLimiter {
+	rl := &RateLimiter{
+		clients:  make(map[string]*clientLimiter),
+		rate:     rate,
+		interval: interval,
+		cleanup:  interval * 10,
+		stopCh:   make(chan struct{}),
+	}
+	go rl.cleanupLoop()
+	return rl
+}
+
+// Close gracefully shuts down the rate limiter cleanup loop
+func (rl *RateLimiter) Close() {
+	rl.stopOnce.Do(func() {
+		close(rl.stopCh)
+	})
+}
+
+func (rl *RateLimiter) cleanupLoop() {
+	ticker := time.NewTicker(rl.cleanup)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-rl.stopCh:
+			return
+		case <-ticker.C:
+			rl.mu.Lock()
+			now := time.Now()
+			for ip, client := range rl.clients {
+				if now.Sub(client.lastCheck) > rl.cleanup {
+					delete(rl.clients, ip)
+				}
+			}
+			rl.mu.Unlock()
+		}
+	}
+}
+
+// Allow checks if a request from the given IP should be allowed
+func (rl *RateLimiter) Allow(ip string) bool {
+	rl.mu.Lock()
+	defer rl.mu.Unlock()
+
+	now := time.Now()
+	client, exists := rl.clients[ip]
+	if !exists {
+		rl.clients[ip] = &clientLimiter{
+			tokens:    rl.rate - 1,
+			lastCheck: now,
+		}
+		return true
+	}
+
+	// Refill tokens based on elapsed time
+	elapsed := now.Sub(client.lastCheck)
+	refill := int(elapsed/rl.interval) * rl.rate
+	client.tokens = min(client.tokens+refill, rl.rate)
+	client.lastCheck = now
+
+	if client.tokens > 0 {
+		client.tokens--
+		return true
+	}
+	return false
+}
